+++ Engine/Engine/time.h
@@ -3,6 +3,6 @@
 
 namespace TIME
 {
-	static constexpr const double SECONDS_MULTIPLYER = 1000000.0;
-	static constexpr const double MILISECONDS_MULTIPLYER = 1000.0;
+	static constexpr const unsigned int SECONDS_MULTIPLYER = 1000000;
+	static constexpr const unsigned int MILISECONDS_MULTIPLYER = 1000;
 }
